+++ src/components/recipe/SwipeCompletion.tsx
